LayoutTests/fast/frames/seamless/seamless-quirks-expected.txt

PASS iframe.seamless is true
PASS iframe.contentDocument.compatMode is "BackCompat"
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
